works? but also doesnt

+++ src/main/java/org/geyser/extension/cordslice/Random/PositionSlicer.java
@@ -31,7 +31,7 @@ public class PositionSlicer {
         }
     }
 
-    public static Vector3i getSlice(Vector3f pos) {
+    public static Vector3i getSlice(Vector3d pos) {
         int sliceX = (int) Math.floor(pos.getX() / SLICE_SIZE);
         int sliceY = (int) Math.floor(pos.getY() / SLICE_SIZE);
         int sliceZ = (int) Math.floor(pos.getZ() / SLICE_SIZE);
@@ -39,7 +39,7 @@ public class PositionSlicer {
         return Vector3i.from(sliceX, sliceY, sliceZ);
     }
 
-    public static SlicedPosition getSlicedPositionFromSlice(Vector3f pos, Vector3i sliced) {
+    public static SlicedPosition getSlicedPositionFromSlice(Vector3d pos, Vector3i sliced) {
         double localX = pos.getX() - (sliced.getX() * SLICE_SIZE);
         double localY = pos.getY() - (sliced.getY() * SLICE_SIZE);
         double localZ = pos.getZ() - (sliced.getZ() * SLICE_SIZE);
@@ -47,7 +47,7 @@ public class PositionSlicer {
         return new SlicedPosition(localX, localY, localZ, sliced.getX(), sliced.getY(), sliced.getZ());
     }
 
-    public static SlicedPosition getSlicedPosition(Vector3f pos) {
+    public static SlicedPosition getSlicedPosition(Vector3d pos) {
         Vector3i sliced = getSlice(pos);
 
         return getSlicedPositionFromSlice(pos, sliced);
